rename: NewDashboard -> NewDashBoardModal 이름 변경
